requirements.txt
alembic==1.7.5
Flask==2.0.2
gunicorn==20.1.0
python-dotenv==0.19.2
SQLAlchemy==1.4.29
